% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleaning}
\alias{cleaning}
\title{Function which produces a vector to be used in string cleaning from scraping there are a lot of hard coded values in here and will need to be adapted for the weird nuances.}
\usage{
cleaning()
}
\description{
Function which produces a vector to be used in string cleaning from scraping there are a lot of hard coded values in here and will need to be adapted for the weird nuances.
}
