% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genius_lyrics.R
\name{genius_lyrics}
\alias{genius_lyrics}
\title{Retrieve song lyrics from Genius.com}
\usage{
genius_lyrics(artist = NULL, song = NULL)
}
\arguments{
\item{artist}{The quoted name of the artist. Spelling matters, capitalization does not.}

\item{song}{The quoted name of the song. Spelling matters, capitalization does not.}
}
\description{
Retrieve the lyrics of a song with supplied artist and song name.
}
\examples{
genius_lyrics(artist = "Margaret Glaspy", song = "Memory Street")
genius_lyrics(artist = "Kendrick Lamar", song = "Money Trees")
genius_lyrics("JMSN", "Drinkin'")

}
