% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{possible_lyrics}
\alias{possible_lyrics}
\title{Form of genius_lyrics that can handle errors}
\usage{
possible_lyrics(...)
}
\arguments{
\item{...}{arguments that would be passed to `genius_lyrics()`}
}
\description{
Form of genius_lyrics that can handle errors
}
