% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{possible_url}
\alias{possible_url}
\title{Form of genius_url that can handle errors}
\usage{
possible_url(...)
}
\arguments{
\item{...}{arguments that would be passed to `genius_url()`}
}
\description{
Form of genius_url that can handle errors
}
